refactor contentservice wip
