+++ src/SFA.DAS.FindEmploymentSchemes.Contentful/Services/Roots/SchemeLengthFilterService.cs
@@ -1,0 +1,18 @@
+﻿using Contentful.Core.Models;
+using SFA.DAS.FindEmploymentSchemes.Contentful.Services.Roots.Base;
+
+namespace SFA.DAS.FindEmploymentSchemes.Contentful.Services.Roots
+{
+    public class SchemeLengthFilterService : FilterService
+    {
+        private const string Name = "schemeLength";
+        private const string Description = "Length of scheme?";
+        private const string ContentfulTypeName = "schemeLengthFilter";
+        public static string Prefix => "scheme-length";
+
+        public SchemeLengthFilterService(HtmlRenderer htmlRenderer)
+            : base(htmlRenderer, Name, Description, ContentfulTypeName, Prefix)
+        {
+        }
+    }
+}
