+++ src/SFA.DAS.FindEmploymentSchemes.Contentful/Services/Roots/SchemeService.cs
@@ -4,71 +4,19 @@ using Contentful.Core.Models;
 using Contentful.Core.Search;
 using Contentful.Core;
 using System.Threading.Tasks;
-using ContentScheme = SFA.DAS.FindEmploymentSchemes.Contentful.Model.Content.Scheme;
+using SFA.DAS.FindEmploymentSchemes.Contentful.Model.Content;
+using SFA.DAS.FindEmploymentSchemes.Contentful.Services.Roots.Base;
 using ApiScheme = SFA.DAS.FindEmploymentSchemes.Contentful.Model.Api.Scheme;
-using Microsoft.AspNetCore.Html;
-using System;
 
 namespace SFA.DAS.FindEmploymentSchemes.Contentful.Services.Roots
 {
-    public class ContentRootService
-    {
-        private readonly HtmlRenderer _htmlRenderer;
-
-        public ContentRootService(HtmlRenderer htmlRenderer)
-        {
-            _htmlRenderer = htmlRenderer;
-        }
-
-        //this one might not belong here
-        protected static string ToFilterAspectId(Model.Api.IFilter filter, string filterPrefix)
-        {
-            return $"{filterPrefix}--{Slugify(filter.Name)}";
-        }
-
-        protected static string Slugify(string? name)
-        {
-            if (name == null)
-                throw new ArgumentNullException(nameof(name));
-
-            return name.ToLower().Replace(' ', '-');
-        }
-
-        protected async Task<HtmlString?> ToHtmlString(Document? document)
-        {
-            if (document == null)
-                return null;
-
-            string html = await _htmlRenderer.ToHtml(document);
-
-            return ToNormalisedHtmlString(html);
-        }
-
-        /// <remarks>
-        /// Should be private, but Contentful's .net library is not very test friendly (HtmlRenderer.ToHtml can't be mocked).
-        /// We'd have to introduce a level of indirection to test this, were it private.
-        /// </remarks>
-        public static HtmlString ToNormalisedHtmlString(string html)
-        {
-            // replace left/right quotation marks with regular old quotation marks
-            html = html.Replace('“', '"').Replace('”', '"');
-
-            // sometimes contentful uses a \r and sometimes a \r\n - nice!
-            // we could strip these out instead
-            html = html.Replace("\r\n", "\r");
-            html = html.Replace("\r", "\r\n");
-
-            return new HtmlString(html);
-        }
-    }
-
     public class SchemeService : ContentRootService
     {
         public SchemeService(HtmlRenderer htmlRenderer) : base(htmlRenderer)
         {
         }
 
-        private async Task<IEnumerable<ContentScheme>> Get(IContentfulClient contentfulClient)
+        private async Task<IEnumerable<Scheme>> Get(IContentfulClient contentfulClient)
         {
             ContentfulCollection<ApiScheme> apiData = await GetSchemesApi(contentfulClient);
             return await SchemesToContent(apiData, contentfulClient.IsPreviewClient);
@@ -81,28 +29,28 @@ namespace SFA.DAS.FindEmploymentSchemes.Contentful.Services.Roots
         }
 
         //todo: IsValid method (in api scheme?)
-        private Task<ContentScheme[]> SchemesToContent(ContentfulCollection<ApiScheme> schemes, bool includeInvalidContent = false)
+        private Task<Scheme[]> SchemesToContent(ContentfulCollection<ApiScheme> schemes, bool includeInvalidContent = false)
         {
             return Task.WhenAll(schemes.Where(s => includeInvalidContent || (s != null && s?.Name != null && s?.ShortDescription != null && s?.ShortBenefits != null && s?.ShortTime != null && s?.Size != null && s?.Url != null))
                 .OrderByDescending(s => s.Size)
                 .Select(ToContent));
         }
 
-        private async Task<ContentScheme> ToContent(ApiScheme apiScheme)
+        private async Task<Scheme> ToContent(ApiScheme apiScheme)
         {
-            IEnumerable<Model.Content.CaseStudy> caseStudies = Enumerable.Empty<Model.Content.CaseStudy>();
+            IEnumerable<CaseStudy> caseStudies = Enumerable.Empty<CaseStudy>();
             if (apiScheme.CaseStudyReferences != null)
             {
                 caseStudies = await Task.WhenAll(apiScheme.CaseStudyReferences.Select(ToContent));
             }
 
-            IEnumerable<Model.Content.SubScheme> subSchemes = Enumerable.Empty<Model.Content.SubScheme>();
+            IEnumerable<SubScheme> subSchemes = Enumerable.Empty<SubScheme>();
             if (apiScheme.SubSchemes != null)
             {
                 subSchemes = await Task.WhenAll(apiScheme.SubSchemes.Select(ToContent));
             }
 
-            return new ContentScheme(
+            return new Scheme(
                 apiScheme.Name!,
                 (await ToHtmlString(apiScheme.ShortDescription))!,
                 (await ToHtmlString(apiScheme.ShortCost))!,
@@ -111,7 +59,7 @@ namespace SFA.DAS.FindEmploymentSchemes.Contentful.Services.Roots
                 apiScheme.Url!,
                 apiScheme.Size,
                 (apiScheme.PayFilterAspects?.Select(f => ToFilterAspectId(f, PayFilterPrefix)) ?? Enumerable.Empty<string>())
-                    .Concat(apiScheme.MotivationsFilterAspects?.Select(f => ToFilterAspectId(f, MotivationsFilterPrefix)) ?? Enumerable.Empty<string>())
+                    .Concat(apiScheme.MotivationsFilterAspects?.Select(f => ToFilterAspectId(f, MotivationFilterService.Prefix)) ?? Enumerable.Empty<string>())
                     .Concat(apiScheme.SchemeLengthFilterAspects?.Select(f => ToFilterAspectId(f, SchemeLengthFilterPrefix)) ?? Enumerable.Empty<string>()),
                 caseStudies,
                 await ToHtmlString(apiScheme.CaseStudies),
@@ -126,17 +74,17 @@ namespace SFA.DAS.FindEmploymentSchemes.Contentful.Services.Roots
                 subSchemes);
         }
 
-        private async Task<Model.Content.CaseStudy> ToContent(Model.Api.CaseStudy apiCaseStudy)
+        private async Task<CaseStudy> ToContent(Model.Api.CaseStudy apiCaseStudy)
         {
             //todo: check all are mandatory in contentful
-            return new Model.Content.CaseStudy(
+            return new CaseStudy(
                 apiCaseStudy.Name!,
                 apiCaseStudy.DisplayTitle!,
                 (await ToHtmlString(apiCaseStudy.Description))!);
         }
-        private async Task<Model.Content.SubScheme> ToContent(Model.Api.SubScheme apiSubScheme)
+        private async Task<SubScheme> ToContent(Model.Api.SubScheme apiSubScheme)
         {
-            return new Model.Content.SubScheme(
+            return new SubScheme(
                 apiSubScheme.Title!,
                 await ToHtmlString(apiSubScheme.Summary),
                 (await ToHtmlString(apiSubScheme.Content))!);
